# backend/app/api/routes/behavior.py
"""Behavior routes (sez. 9): capture init/complete, event read, feedback."""

from __future__ import annotations

from fastapi import APIRouter, Depends

from app.api.deps import IdempotencyDep, StateDep, UserIdDep, rate_limit
from app.contracts.api import (
    BehaviorCaptureInitRequest,
    BehaviorCaptureInitResponse,
    BehaviorContextUpdateRequest,
    BehaviorEventOut,
    BehaviorFeedbackRequest,
    BehaviorFeedbackResponse,
    CaptureCompleteResponse,
)
from app.contracts.taxonomy import INTERPRETATION_SCHEMA_VERSION, FeedbackValue
from app.domains import behavior as behavior_domain
from app.domains import behavior_db, idempotency_db, lifestyle, lifestyle_db
from app.domains.context_bucket import resolve_context_bucket
from app.domains.models import BehaviorEventRec
from app.knowledge.models import AdviceOutcomeValue

router = APIRouter()


def event_out(
    event: BehaviorEventRec,
    feedback: FeedbackValue | None = None,
    advice_outcome: AdviceOutcomeValue | None = None,
) -> BehaviorEventOut:
    interp = event.interpretation_json or {}
    consumer = interp.get("consumer") or {}
    return BehaviorEventOut(
        id=event.id,
        dog_id=event.dog_id,
        status=event.status,
        schema_version=interp.get("schema_version", INTERPRETATION_SCHEMA_VERSION),
        primary_intent=event.primary_intent,
        confidence_band=event.confidence_band,
        summary=event.summary,
        alternatives=interp.get("alternatives", []),
        evidence=interp.get("evidence", []),
        safety_flags=interp.get("safety_flags", []),
        needs_context=interp.get("needs_context", False),
        context_question=interp.get("context_question"),
        policy_version=event.policy_version,
        taxonomy_version=event.taxonomy_version,
        feedback=feedback,
        advice=event.advice_json or interp.get("advice"),
        advice_outcome=advice_outcome,
        consumer_headline=consumer.get("consumer_headline"),
        baseline_comparison=consumer.get("baseline_comparison"),
        baseline_note=consumer.get("baseline_note"),
        recommended_next_step=consumer.get("recommended_next_step"),
        what_to_watch=consumer.get("what_to_watch"),
        safety=consumer.get("safety"),
        personal_memory_used=consumer.get("personal_memory_used")
        or interp.get("personal_memory_used")
        or [],
        context_bucket=interp.get("context_bucket"),
        created_at=event.created_at,
        completed_at=event.completed_at,
    )


async def _record_guard(state: StateDep, guard: IdempotencyDep, body: dict) -> None:
    guard.record(body)
    if state.engine is not None and guard._scope:
        await idempotency_db.record(
            state.engine,
            scope=guard._scope,
            body=body,
            payload_hash=guard._payload_hash,
        )


@router.post("/behavior/captures/init", response_model=BehaviorCaptureInitResponse)
async def init_capture(
    payload: BehaviorCaptureInitRequest,
    state: StateDep,
    user_id: UserIdDep,
    guard: IdempotencyDep,
    _limiter: None = Depends(rate_limit("behavior.init", limit=30)),
) -> BehaviorCaptureInitResponse:
    if cached := guard.lookup():
        return BehaviorCaptureInitResponse.model_validate(cached)
    if state.engine is not None:
        lifestyle_out = await lifestyle_db.get_lifestyle(
            state.engine, user_id, payload.dog_id
        )
    else:
        lifestyle_out = lifestyle.get_lifestyle(
            state.store, user_id, payload.dog_id
        )
    resolved_bucket = resolve_context_bucket(
        payload.context_bucket, lifestyle=lifestyle_out.model_dump()
    )
    if resolved_bucket != payload.context_bucket:
        payload = payload.model_copy(update={"context_bucket": resolved_bucket})
    if state.engine is not None:
        capture, event, url, expires, reserved = await behavior_db.init_capture(
            state.engine,
            settings=state.settings,
            storage=state.storage,
            user_id=user_id,
            payload=payload,
        )
    else:
        capture, event, url, expires, reserved = await behavior_domain.init_capture(
            state.store,
            settings=state.settings,
            storage=state.storage,
            user_id=user_id,
            payload=payload,
        )
    resp = BehaviorCaptureInitResponse(
        capture_id=capture.id,
        event_id=event.id,
        status=event.status,
        upload={"url": url, "storage_path": capture.storage_path, "expires_at": expires},
        quota_reserved=reserved,
    )
    await _record_guard(state, guard, resp.model_dump(mode="json"))
    return resp


@router.post("/behavior/captures/{capture_id}/complete", response_model=CaptureCompleteResponse)
async def complete_capture(
    capture_id: str,
    state: StateDep,
    user_id: UserIdDep,
    guard: IdempotencyDep,
) -> CaptureCompleteResponse:
    if cached := guard.lookup():
        return CaptureCompleteResponse.model_validate(cached)
    if state.engine is not None:
        event = await behavior_db.complete_capture(
            state.engine,
            settings=state.settings,
            storage=state.storage,
            queue=state.queue,
            user_id=user_id,
            capture_id=capture_id,
        )
    else:
        event = await behavior_domain.complete_capture(
            state.store,
            settings=state.settings,
            storage=state.storage,
            queue=state.queue,
            user_id=user_id,
            capture_id=capture_id,
        )
    resp = CaptureCompleteResponse(capture_id=capture_id, event_id=event.id, status=event.status)
    await _record_guard(state, guard, resp.model_dump(mode="json"))
    return resp


@router.get("/behavior/events/{event_id}", response_model=BehaviorEventOut)
async def get_behavior_event(event_id: str, state: StateDep, user_id: UserIdDep) -> BehaviorEventOut:
    if state.engine is not None:
        event = await behavior_db.get_event(state.engine, user_id=user_id, event_id=event_id)
        feedback = await behavior_db.get_feedback_value(
            state.engine, user_id=user_id, event_id=event_id
        )
        advice_outcome = await lifestyle_db.get_latest_advice_outcome(
            state.engine, user_id=user_id, event_id=event_id
        )
    else:
        event = behavior_domain.get_event(state.store, user_id=user_id, event_id=event_id)
        rec = state.store.behavior_feedback.get(event_id)
        feedback = rec.value if rec is not None and rec.user_id == user_id else None
        advice_outcome = lifestyle.get_latest_advice_outcome(
            state.store, user_id=user_id, event_id=event_id
        )
    return event_out(event, feedback, advice_outcome)


@router.post(
    "/behavior/events/{event_id}/context",
    response_model=BehaviorEventOut,
)
async def update_behavior_context(
    event_id: str,
    payload: BehaviorContextUpdateRequest,
    state: StateDep,
    user_id: UserIdDep,
    _limiter: None = Depends(rate_limit("behavior.context", limit=20)),
) -> BehaviorEventOut:
    """Save one owner answer and refine without observing the video again."""
    if state.engine is not None:
        event = await behavior_db.get_event(
            state.engine, user_id=user_id, event_id=event_id
        )
    else:
        event = behavior_domain.get_event(
            state.store, user_id=user_id, event_id=event_id
        )

    # Local import keeps the public route independent from worker startup.
    from app.worker.handlers import refine_behavior_event_context

    event = await refine_behavior_event_context(
        state,
        event=event,
        context_bucket=payload.context_bucket,
    )
    return event_out(event)


@router.post("/behavior/events/{event_id}/feedback", response_model=BehaviorFeedbackResponse)
async def post_feedback(
    event_id: str,
    payload: BehaviorFeedbackRequest,
    state: StateDep,
    user_id: UserIdDep,
    guard: IdempotencyDep,
    _limiter: None = Depends(rate_limit("behavior.feedback", limit=60)),
) -> BehaviorFeedbackResponse:
    if cached := guard.lookup():
        return BehaviorFeedbackResponse.model_validate(cached)
    if state.engine is not None:
        rec = await behavior_db.record_feedback(
            state.engine, user_id=user_id, event_id=event_id, payload=payload
        )
    else:
        rec = behavior_domain.record_feedback(
            state.store, user_id=user_id, event_id=event_id, payload=payload
        )
    resp = BehaviorFeedbackResponse(event_id=event_id, value=rec.value)
    await _record_guard(state, guard, resp.model_dump(mode="json"))
    return resp