# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dogly API",
    "description": "Public API V1 (Spec V1 sez. 9). OpenAPI is the mobile client contract.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/behavior/captures/init": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Init Capture",
        "operationId": "init_capture_v1_behavior_captures_init_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehaviorCaptureInitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehaviorCaptureInitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/behavior/captures/{capture_id}/complete": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Complete Capture",
        "operationId": "complete_capture_v1_behavior_captures__capture_id__complete_post",
        "parameters": [
          {
            "name": "capture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capture Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/behavior/events/{event_id}": {
      "get": {
        "tags": [
          "behavior"
        ],
        "summary": "Get Behavior Event",
        "operationId": "get_behavior_event_v1_behavior_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehaviorEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/behavior/events/{event_id}/context": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Update Behavior Context",
        "description": "Save one owner answer and refine without observing the video again.",
        "operationId": "update_behavior_context_v1_behavior_events__event_id__context_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehaviorContextUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehaviorEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/behavior/events/{event_id}/feedback": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Post Feedback",
        "operationId": "post_feedback_v1_behavior_events__event_id__feedback_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehaviorFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehaviorFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdviceItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "follow_up": {
            "type": "string",
            "title": "Follow Up"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "risk": {
            "type": "string",
            "title": "Risk"
          }
        },
        "type": "object",
        "required": [
          "code",
          "category",
          "action",
          "rationale",
          "follow_up",
          "source_ids",
          "risk"
        ],
        "title": "AdviceItem"
      },
...
      "AdviceOutcomeValue": {
        "type": "string",
        "enum": [
          "HELPED",
          "DID_NOT_HELP",
          "UNKNOWN",
          "NOT_TRIED"
        ],
        "title": "AdviceOutcomeValue"
      },
      "AlternativeIntent": {
        "properties": {
          "intent": {
            "$ref": "#/components/schemas/IntentCode"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "intent",
          "rationale"
        ],
        "title": "AlternativeIntent"
      },
...
      "BehaviorCaptureInitRequest": {
        "properties": {
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "client_request_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Client Request Id"
          },
          "duration_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Duration Ms"
          },
          "has_audio": {
            "type": "boolean",
            "title": "Has Audio",
            "default": true
          },
          "bytes": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Bytes"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "video/mp4",
              "video/quicktime",
              "video/webm"
            ],
            "title": "Content Type",
            "default": "video/mp4"
          },
          "context_bucket": {
            "$ref": "#/components/schemas/ContextBucket",
            "default": "UNKNOWN"
          }
        },
        "type": "object",
        "required": [
          "dog_id",
          "client_request_id",
          "duration_ms",
          "bytes"
        ],
        "title": "BehaviorCaptureInitRequest"
      },
      "BehaviorCaptureInitResponse": {
        "properties": {
          "capture_id": {
            "type": "string",
            "title": "Capture Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status": {
            "$ref": "#/components/schemas/BehaviorEventStatus"
          },
          "upload": {
            "$ref": "#/components/schemas/SignedUpload"
          },
          "quota_reserved": {
            "type": "boolean",
            "title": "Quota Reserved",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "capture_id",
          "event_id",
          "status",
          "upload"
        ],
        "title": "BehaviorCaptureInitResponse"
      },
      "BehaviorContextUpdateRequest": {
        "properties": {
          "context_bucket": {
            "$ref": "#/components/schemas/ContextBucket"
          }
        },
        "type": "object",
        "required": [
          "context_bucket"
        ],
        "title": "BehaviorContextUpdateRequest",
        "description": "One owner answer that can materially refine a behavior result."
      },
      "BehaviorEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "status": {
            "$ref": "#/components/schemas/BehaviorEventStatus"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "primary_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IntentCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativeIntent"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "safety_flags": {
            "items": {
              "$ref": "#/components/schemas/SafetyFlag"
            },
            "type": "array",
            "title": "Safety Flags"
          },
          "needs_context": {
            "type": "boolean",
            "title": "Needs Context",
            "default": false
          },
          "context_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Question"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "taxonomy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxonomy Version"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "advice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdviceItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "advice_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdviceOutcomeValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "consumer_headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumer Headline"
          },
          "baseline_comparison": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Comparison"
          },
          "baseline_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Note"
          },
          "recommended_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Next Step"
          },
          "what_to_watch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "What To Watch"
          },
          "safety": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety"
          },
          "personal_memory_used": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Personal Memory Used"
          },
          "context_bucket": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextBucket"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dog_id",
          "status",
          "schema_version",
          "created_at"
        ],
        "title": "BehaviorEventOut"
      },
      "BehaviorEventStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "UPLOADING",
          "QUEUED",
          "OBSERVING",
          "INTERPRETING",
          "COMPLETED",
          "REJECTED_QUALITY",
          "FAILED_RETRYABLE",
          "FAILED_TERMINAL",
          "CANCELLED"
        ],
        "title": "BehaviorEventStatus",
        "description": "Behavior analysis state machine (sez. 7.2 / 33.1)."
      },
      "BehaviorFeedbackRequest": {
        "properties": {
          "value": {
            "$ref": "#/components/schemas/FeedbackValue"
          },
          "correction_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IntentCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "corrected_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextBucket"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "BehaviorFeedbackRequest"
      },
      "BehaviorFeedbackResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "value": {
            "$ref": "#/components/schemas/FeedbackValue"
          },
          "recorded": {
            "type": "boolean",
            "title": "Recorded",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "value"
        ],
        "title": "BehaviorFeedbackResponse"
      },
      "CaptureCompleteResponse": {
        "properties": {
          "capture_id": {
            "type": "string",
            "title": "Capture Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status": {
            "$ref": "#/components/schemas/BehaviorEventStatus"
          }
        },
        "type": "object",
        "required": [
          "capture_id",
          "event_id",
          "status"
        ],
        "title": "CaptureCompleteResponse"
      },
...
      "ConfidenceBand": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "ConfidenceBand"
      },
      "ContextBucket": {
        "type": "string",
        "enum": [
          "HOME",
          "OUTDOORS",
          "WALK",
          "PLAY",
          "FEEDING",
          "DOOR_EXIT",
          "REST",
          "STRANGER",
          "OTHER_DOG",
          "VEHICLE",
          "HANDLING",
          "UNKNOWN"
        ],
        "title": "ContextBucket",
        "description": "Context buckets V0 (sez. 33.7)."
      },
...
      "EvidenceItem": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/EvidenceSource"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "description"
        ],
        "title": "EvidenceItem"
      },
      "EvidenceSource": {
        "type": "string",
        "enum": [
          "observation",
          "context",
          "personal_pattern",
          "scientific_kb",
          "life_stage",
          "lifestyle_baseline"
        ],
        "title": "EvidenceSource"
      },
...
      "FeedbackValue": {
        "type": "string",
        "enum": [
          "YES",
          "NO",
          "UNKNOWN"
        ],
        "title": "FeedbackValue"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntentCode": {
        "type": "string",
        "enum": [
          "PLAY_INTERACTION",
          "ATTENTION_REQUEST",
          "OUTSIDE_REQUEST",
          "ALERT_VIGILANCE",
          "DISCOMFORT_AVOIDANCE",
          "FEAR_INSECURITY",
          "HIGH_AROUSAL",
          "FRUSTRATION",
          "RELAX_REST",
          "RESOURCE_TENSION",
          "AMBIGUOUS",
          "INSUFFICIENT"
        ],
        "title": "IntentCode",
        "description": "Initial closed intent taxonomy (sez. 16.2).\n\nClosed allowlist: provider output outside this set is schema-invalid."
      },
...
      "SafetyFlag": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "info"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "SafetyFlag",
        "description": "Structured flags consumed by the deterministic copy layer (sez. 16.3).\nGenerated text may never downgrade a safety flag (sez. 19.3)."
      },
...
      "SignedUpload": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "storage_path": {
            "type": "string",
            "title": "Storage Path"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "url",
          "storage_path",
          "expires_at"
        ],
        "title": "SignedUpload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}